+++ src/_pytask/build.py
@@ -233,12 +233,10 @@ def build(  # noqa: C901, PLR0912, PLR0913, PLR0915
         session = Session.from_config(config_)
 
     except (ConfigurationError, Exception):
-        exc_info = sys.exc_info()
-        exc_info = remove_internal_traceback_frames_from_exc_info(exc_info)
+        exc_info = remove_internal_traceback_frames_from_exc_info(sys.exc_info())
         traceback = Traceback.from_exception(*exc_info)
         console.print(traceback)
-        session = Session({}, None)
-        session.exit_code = ExitCode.CONFIGURATION_FAILED
+        session = Session(exit_code=ExitCode.CONFIGURATION_FAILED)
 
     else:
         try:
@@ -257,8 +255,7 @@ def build(  # noqa: C901, PLR0912, PLR0913, PLR0915
             session.exit_code = ExitCode.FAILED
 
         except Exception:  # noqa: BLE001
-            exc_info = sys.exc_info()
-            exc_info = remove_internal_traceback_frames_from_exc_info(exc_info)
+            exc_info = remove_internal_traceback_frames_from_exc_info(sys.exc_info())
             traceback = Traceback.from_exception(*exc_info)
             console.print(traceback)
             session.exit_code = ExitCode.FAILED