+++ docs/source/changes.md
@@ -55,6 +55,7 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 - {pull}`437` fixes the detection of task functions and publishes
   {func}`pytask.is_task_function`.
 - {pull}`438` clarifies some types.
+- {pull}`440` refines more types.
 
 ## 0.3.2 - 2023-06-07